+++ src/main/java/mcp/mobius/opis/data/profilers/ProfilerRenderBlock.java
@@ -5,8 +5,6 @@ import java.util.HashMap;
 import mcp.mobius.opis.data.holders.basetypes.CoordinatesBlock;
 import mcp.mobius.opis.data.profilers.Clock.IClock;
 import net.minecraft.client.Minecraft;
-import net.minecraft.world.World;
-
 import org.apache.commons.math3.stat.descriptive.DescriptiveStatistics;
 
 public class ProfilerRenderBlock extends ProfilerAbstract {
@@ -32,7 +30,7 @@ public class ProfilerRenderBlock extends ProfilerAbstract {
 		CoordinatesBlock coord = new CoordinatesBlock(Minecraft.getMinecraft().theWorld.provider.dimensionId, (Integer)key2, (Integer)key3, (Integer)key4);
 		
 		clock.stop();
-		data.get(coord).addValue((double)clock.getDelta());		
+		data.get(coord).addValue(clock.getDelta());		
 	}		
 	
 }
