+++ src/main/java/mcp/mobius/opis/data/profilers/ProfilerRenderEntity.java
@@ -32,7 +32,7 @@ public class ProfilerRenderEntity extends ProfilerAbstract {
 		clock.stop();
 		try{
 			//data.get((Entity)key).addValue((double)clock.getDelta());
-			data.get(key).addValue((double)clock.getDelta());
+			data.get(key).addValue(clock.getDelta());
 		} catch (Exception e) {
 			modOpis.log.warning(String.format("Error while profiling entity %s\n", key));
 		}
