+++ src/main/java/mcp/mobius/opis/data/profilers/ProfilerTick.java
@@ -1,7 +1,5 @@
 package mcp.mobius.opis.data.profilers;
 
-import java.util.ArrayList;
-
 import mcp.mobius.opis.data.profilers.Clock.IClock;
 
 import org.apache.commons.math3.stat.descriptive.DescriptiveStatistics;
@@ -25,7 +23,7 @@ public class ProfilerTick extends ProfilerAbstract {
 	@Override
 	public void stop() {
 		this.clock.stop();
-		this.data.addValue((double)clock.getDelta());
+		this.data.addValue(clock.getDelta());
 	}
 	
 }
